requirements.txt
runpod==1.3.4
python-dotenv==1.0.0
llama-cpp-python==0.1.77
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git